Remerge Cliff's changes for .babelrc
