+++ index.js
@@ -525,136 +525,7 @@ function bundleJs(moduleToBundle, as) {
         }
 
         exports.defineBundleTask(bundleTaskName, function() {
-            if (!bundle.bundleInDir) {
-                var adjunctBase = setAdjunctInDir(bundle);
-                logger.logInfo('Javascript bundle "' + bundle.as + '" will be available in Jenkins as adjunct "' + adjunctBase + '.' + bundle.as + '".')
-            }
-
-            // Add all global mappings.
-            if (bundle.useGlobalModuleMappings === true) {
-                for (var i = 0; i < globalModuleMappingArgs.length; i++) {
-                    bundle._withExternalModuleMapping(globalModuleMappingArgs[i]);
-                }
-            }
-
-            var bundleTo = bundle.bundleInDir;
-
-            if (!applyImports) {
-                bundleTo += '/no_imports';
-            }
-
-            // Only process LESS when generating the bundle containing imports. If using the "no_imports" bundle, you
-            // need to take care of adding the CSS yourself.
-            if (applyImports && bundle.lessSrcPath) {
-                var lessBundleTo = bundleTo;
-
-                if (bundle.lessTargetDir) {
-                    lessBundleTo = bundle.lessTargetDir;
-                }
-
-                less(bundle.lessSrcPath, lessBundleTo);
-            }
-
-            var fileToBundle = bundle.bundleModule;
-            if (bundle.bundleDependencyModule) {
-                // Lets generate a temp file containing the module require.
-                if (!fs.existsSync('target')) {
-                    fs.mkdirSync('target');
-                }
-                fileToBundle = 'target/' + bundle.bundleOutputFile;
-                fs.writeFileSync(fileToBundle, "module.exports = require('" + bundle.module + "');");
-            }
-
-            var fullPaths = args.isArgvSpecified('--full-paths');
-            
-            var browserifyConfig = {
-                entries: [fileToBundle],
-                extensions: ['.js', '.es6', '.jsx', '.hbs'],
-                cache: {},
-                packageCache: {},
-                fullPaths: fullPaths
-            };
-            if (bundle.minifyBundle === true) {
-                browserifyConfig.debug = true;
-            }
-            var bundler = browserify(browserifyConfig);
-
-            var hasJSX = paths.hasSourceFiles('jsx');
-            var hasES6 = paths.hasSourceFiles('es6');
-            var hasBabelRc = fs.existsSync('.babelrc');
-
-            if (langConfig.ecmaVersion === 6 || hasJSX || hasES6 || hasBabelRc) {
-                var babelify = require('babelify');
-                var presets = [];
-                var plugins = [];
-
-                if (hasBabelRc) {
-                    logger.logInfo("Will use babel config from .babelrc");
-                }
-                else if (hasJSX) {
-                    presets.push('react');
-                    dependencies.warnOnMissingDependency('babel-preset-react', 'You have JSX sources in this project. Transpiling these will require the "babel-preset-react" package.');
-                    presets.push('es2015');
-                    dependencies.warnOnMissingDependency('babel-preset-es2015', 'You have JSX/ES6 sources in this project. Transpiling these will require the "babel-preset-es2015" package.');
-                } else {
-                    presets.push('es2015');
-                    dependencies.warnOnMissingDependency('babel-preset-es2015', 'You have ES6 sources in this project. Transpiling these will require the "babel-preset-es2015" package.');
-                }
-
-                var babelConfig = {};
-
-                // if no .babelrc was found, configure babel with the default presets and plugins from above
-                if (!hasBabelRc) {
-                    babelConfig.presets = presets;
-                    babelConfig.plugins = plugins;
-                }
-
-                // if .babelrc was found, an empty config object must be passed in order for .babelrc config to be read automatically
-                bundler.transform(babelify, babelConfig);
-            }
-
-            if (bundle.bundleTransforms) {
-                for (var i = 0; i < bundle.bundleTransforms.length; i++) {
-                    bundler.transform(bundle.bundleTransforms[i]);
-                }
-            }
-
-            if (applyImports) {
-                addModuleMappingTransforms(bundle, bundler);
-            }
-
-            if (bundle.minifyBundle === true) {
-                var sourceMap = bundle.as + '.map.json';
-                bundler.plugin('minifyify', {
-                    map: sourceMap,
-                    output: bundleTo + '/' + sourceMap
-                });
-            }
-
-            for (var i = 0; i < preBundleListeners.length; i++) {
-                preBundleListeners[i].call(bundle, bundler);
-            }
-            
-            // Allow reading of stuff from the filesystem.
-            bundler.transform(require('brfs'));
-            
-            return bundler.bundle()
-                .on('error', function (err) {
-                    logger.logError('Browserify bundle processing error');
-                    if (err) {
-                        logger.logError('\terror: ' + err);
-                    }
-                    if (exports.isRebundle() || exports.isRetest()) {
-                        notifier.notify('bundle:watch failure', 'See console for details.');
-                        // ignore failures if we are running rebundle/retesting.
-                        this.emit('end');
-                    } else {
-                        throw 'Browserify bundle processing error. See above for details.';
-                    }
-                })
-                .pipe(source(bundle.bundleOutputFile))
-                .pipe(gulp.dest(bundleTo))
-                ;
+            return bundlegen.doJSBundle(bundle, applyImports);
         });
     }
